requirements.txt
albumentations==0.4.6
bypy==1.7.12
cached-property==1.5.2
cachetools==4.1.1
Cython==0.29.21
cytoolz==0.9.0.1
decorator==4.4.2
dill==0.3.5.1
h5py==3.1.0
hydra-core==1.1.1
idna==2.10
imagecorruptions==1.1.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
importlib-resources==5.3.0
Jinja2==3.0.3
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
lmdb==1.2.1
lpips==0.1.3
lxml==4.7.1
Markdown==3.2.2
MarkupSafe==2.0.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv==0.6.2
msgpack==0.6.2
msgpack-numpy==0.4.7.1
multiprocess==0.70.13
murmurhash==0.28.0
natsort==7.1.1
networkx==2.5
numpy==1.18.5
nvidia-ml-py==12.535.77
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.1
opencv-python
pandas==1.1.5
pathlib==1.0.1
Pillow==6.2.2
protobuf==3.12.2
pycocotools==12.0
python-dateutil==2.1
pytorch-wavelets
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
rawpy==0.15.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==1.0.2
Scio==0.12.0
scipy==1.2.1
seaborn==0.11.2
Shapely==1.7.1
six==1.15.0
sklearn==0.0.post1
soupsieve==2.3.1
spacy
tb-nightly==2.8.0a20211221
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toolz==0.11.1
torch==1.2.0
torchfile==0.1.0
torchstat==0.0.7
torchvision==0.5.0
tornado==6.1
tqdm==4.51.0
typing-extensions==4.3.0
ujson==4.0.2
urllib3==1.25.10
visdom==0.1.8.9
warmup-scheduler==0.3
wcwidth==0.2.6
websocket-client==1.1.0
Werkzeug==1.0.1
wrapt==1.10.11
yacs==0.1.8
yapf==0.30.0
zipp==3.1.0
